fix(mouse): implement universal 12-bit packed, 16-bit and 8-bit BLE mouse report decoding

+++ src/services/solar_os_mouse.c
@@ -78,7 +78,7 @@ bool solar_os_mouse_is_connected(void)
     return connected;
 }
 
-void solar_os_mouse_process_report(uint8_t buttons, int8_t dx, int8_t dy, int8_t wheel)
+void solar_os_mouse_process_report(uint8_t buttons, int16_t dx, int16_t dy, int8_t wheel)
 {
     const uint32_t now = mouse_millis();
